chore: moved `instantiate_circuit` and `instantiate_empty_circuit` function as methods to MerkleSumTreeCircuit struct
